# services/worker/openapi.yaml
openapi: 3.0.3
info:
  title: Papertrail Worker API
  description: |
    Invoice processing worker for the Papertrail Invoice Bot.

    This service processes invoices through the pipeline:
    1. Download from Telegram
    2. Upload to Cloud Storage
    3. Extract data with LLM Vision (Gemini/OpenAI)
    4. Append to Google Sheets
    5. Send confirmation to Telegram

    ## Authentication
    - Processing endpoints are called by Cloud Tasks with OIDC tokens
    - Health and metrics endpoints are public

    ## Job States
    | Status | Description |
    |--------|-------------|
    | `pending` | Waiting in queue |
    | `processing` | Currently being processed |
    | `pending_retry` | Failed, waiting for Cloud Tasks retry |
    | `processed` | Successfully completed |
    | `failed` | Failed after max retries |
    | `pending_decision` | Duplicate detected, awaiting user input |
  version: "1.0.0"
  contact:
    name: Papertrail Team
servers:
  - url: https://worker-XXXXX.us-central1.run.app
    description: Production
  - url: http://localhost:8081
    description: Local development
...
paths:
...
  /invoice/command:
    post:
      tags:
        - Invoice Generation
      summary: Handle /invoice command
      description: Starts a new invoice generation session
      operationId: handleInvoiceCommand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceCommandPayload"
      responses:
        "200":
          description: Command handled
...
components:
  securitySchemes:
    cloudTasks:
      type: http
      scheme: bearer
      bearerFormat: OIDC
      description: Cloud Tasks OIDC token
  schemas:
...
    InvoiceCommandPayload:
      type: object
      properties:
        type:
          type: string
          enum: [command]
        chatId:
          type: integer
        messageId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        firstName:
          type: string
        text:
          type: string
          description: Full command text
        receivedAt:
          type: string
          format: date-time
...